STRY52090541: Used interface for custom credential resolvers
